+++ frontend/src/components/StorePage/utils/colorsForReviewBox.js
@@ -1,0 +1,10 @@
+const colors = [
+    'pink',
+    'orange',
+    'red',
+    'blue',
+    'yellow',
+    'green',
+]
+
+export default colors
