+++ frontend/src/components/StorePage/utils/iconsForReviewBox.jsx
@@ -1,0 +1,12 @@
+import { FaAppleAlt, FaLemon, FaCarrot, FaLeaf, FaSeedling, FaPepperHot } from "react-icons/fa"
+
+const icons = [
+    <FaAppleAlt style={{ fontSize: '25px', color: 'red' }} />,
+    <FaSeedling style={{ fontSize: '25px', color: 'green' }} />,
+    <FaCarrot style={{ fontSize: '25px', color: 'orange' }} />,
+    <FaLeaf style={{ fontSize: '25px', color: 'green' }} />,
+    <FaPepperHot style={{ fontSize: '25px', color: 'red' }} />,
+    <FaLemon style={{ fontSize: '25px', color: 'yellow' }} />
+]
+
+export default icons
