+++ frontend/src/context/Modal.css
@@ -7,6 +7,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
+  z-index: 1;
 }
 
 #modal-background {
@@ -16,9 +17,11 @@
   left: 0;
   bottom: 0;
   background-color: rgba(0, 0, 0, 0.7);
+  z-index: 1;
 }
 
 #modal-content {
   position: absolute;
   background-color:white;
+  z-index: 1;
 }
